refac: cleaned App.js
